requirements.txt
torch
torchvision
transformers
numpy
PyYaml
tqdm
